rally in openstack instance
this rally works in an openstack instance, definitely works in a local
VM. for 10.1.1.180 cluster, you need to explicitly assign the
shared-internal network to run rally instance, while it is not required
in 10.1.0.2. After a series of trial-and-error, we found out that in
10.1.1.180 cluster, ANY user in ANY tenant/project with an ADMIN role
can launch rally in an EXPLICIT SHARED NON-EXTERNAL network.